Move pending actions to an Actions tab with year/class drill-down (#154)
The Reconcile screen rendered the whole pending-actions list inline, so a
September changeover's hundreds/thousands of tiles made it unresponsive.

- Add an "Actions" nav destination (Home / Reconcile / Actions / Passwords /
  Settings), sharing the one memoized ReconcileServices/ReconcileController.
- Move the pending-actions section (title, Dry-run all / Apply all, the
  action tiles) and the school → grade-year → classroom rollup drill-down out
  of Reconcile onto the Actions tab. Reconcile keeps the linked-overview
  counts, sync/drift controls, duplicate-mail warnings, and the resizable log.
- Actions browses by the existing #119 rollup tree and only loads/builds one
  classroom's actions when the operator drills in (LinkedStore.readClassroom),
  rendering that class's tiles through a lazy SliverList so only on-screen
  tiles build. Active sessions show the interactive per-entry choice/apply
  tiles (joined to the classroom docs by account id); passive sessions show
  the read-only materialized candidate summaries.
- Controller: add classroomPendingEntries / groupPendingEntries (+ situation
  groupings) and totalPendingCount; share the situation-grouping helper.

Tests: new actions_screen_test.dart (drill-down loads one classroom lazily,
sliver list virtualizes, #110 choice/#111 virtualization/#153 diff/#115+#119
passive drill-downs); trimmed reconcile_screen_test.dart to the reconcile-only
surface and asserts the flat list is gone; integration flows updated to browse
actions via the Actions tab.
