+++ account_manager/lib/src/screens/reconcile_screen.dart
@@ -1,9 +1,7 @@
 import 'dart:async';
 
-import 'package:account_actions/account_actions.dart' as actions;
 import 'package:account_core/account_core.dart' as core;
-import 'package:account_state/account_state.dart'
-    show LinkedState, MaterializedAccount, MaterializedGroup, Rollup;
+import 'package:account_state/account_state.dart' show LinkedState;
 import 'package:flutter/material.dart';
 import 'package:plink_design_system/plink_design_system.dart';
 
@@ -105,27 +103,6 @@ class _ReconcileScreenState extends State<ReconcileScreen> {
   }
 }
 
-/// One row of the flattened pending-actions list (#111): a same-situation
-/// bulk header or a single account entry. Flattening the situation → entries
-/// tree into a linear list lets a lazy [SliverList] build only the on-screen
-/// rows, so a September changeover's hundreds/thousands of tiles no longer all
-/// build, lay out, and paint every frame.
-sealed class _PendingRow {
-  const _PendingRow();
-}
-
-class _SituationHeaderRow extends _PendingRow {
-  const _SituationHeaderRow(this.entries);
-
-  final List<PendingAccountEntry> entries;
-}
-
-class _EntryRow extends _PendingRow {
-  const _EntryRow(this.entry);
-
-  final PendingAccountEntry entry;
-}
-
 class _ReconcileBody extends StatelessWidget {
   const _ReconcileBody({required this.controller, required this.log});
 
@@ -175,6 +152,10 @@ class _ReconcileBody extends StatelessWidget {
   static SliverToBoxAdapter _gap(double height) =>
       SliverToBoxAdapter(child: SizedBox(height: height));
 
+  /// Reconcile is now short and responsive (#154): the sync/drift header, the
+  /// status banner, and — once a sync has linked this session — the linked
+  /// overview counts with the duplicate-mail warnings. The pending actions and
+  /// their per-year/per-class drill-down live on the Actions tab.
   List<Widget> _slivers() {
     final linked = controller.linked;
     final slivers = <Widget>[
@@ -183,120 +164,15 @@ class _ReconcileBody extends StatelessWidget {
       _gap(PlinkSpacing.s5),
       _section(_StatusBanner(controller: controller)),
     ];
-    if (controller.selectedClassroom != null) {
-      slivers
-        ..add(_gap(PlinkSpacing.s5))
-        ..add(_section(_ClassroomDetail(controller: controller)));
-    } else if (controller.showingGroups) {
-      slivers
-        ..add(_gap(PlinkSpacing.s5))
-        ..add(_section(_GroupsDetail(controller: controller)));
-    } else if (controller.hasOverview) {
-      slivers
-        ..add(_gap(PlinkSpacing.s5))
-        ..add(_section(_DrillDownSection(controller: controller)));
-    }
     if (linked != null) {
       slivers
         ..add(_gap(PlinkSpacing.s5))
         ..add(
-            _section(_OverviewSection(linked: linked, controller: controller)))
-        ..add(_gap(PlinkSpacing.s5))
-        ..addAll(_actionsSlivers());
+            _section(_OverviewSection(linked: linked, controller: controller)));
     }
-    slivers
-      ..addAll(_resultsSlivers())
-      ..add(_gap(PlinkSpacing.s6));
+    slivers.add(_gap(PlinkSpacing.s6));
     return slivers;
   }
-
-  /// The pending-actions section (#110/#111): the title and dry-run/apply
-  /// buttons stay as adapters, but the situation headers and per-account entry
-  /// tiles render through a lazy [SliverList] so a huge pending set does not
-  /// build every tile up front.
-  List<Widget> _actionsSlivers() {
-    final entries = controller.pendingEntries;
-    if (entries.isEmpty) {
-      return <Widget>[_section(const _PendingActionsHeader(count: 0))];
-    }
-    final rows = _pendingRows(controller.pendingSituations);
-    return <Widget>[
-      _section(_PendingActionsHeader(count: entries.length)),
-      SliverPadding(
-        padding: _hPad,
-        sliver: SliverList.builder(
-          itemCount: rows.length,
-          itemBuilder: (context, index) {
-            final row = rows[index];
-            return switch (row) {
-              _SituationHeaderRow(:final entries) =>
-                _SituationHeader(controller: controller, entries: entries),
-              _EntryRow(:final entry) =>
-                _PendingEntryTile(controller: controller, entry: entry),
-            };
-          },
-        ),
-      ),
-      _section(_PendingActionsFooter(controller: controller)),
-    ];
-  }
-
-  static List<_PendingRow> _pendingRows(
-    List<List<PendingAccountEntry>> situations,
-  ) {
-    final rows = <_PendingRow>[];
-    for (final subset in situations) {
-      if (subset.length > 1) rows.add(_SituationHeaderRow(subset));
-      for (final entry in subset) {
-        rows.add(_EntryRow(entry));
-      }
-    }
-    return rows;
-  }
-
-  List<Widget> _resultsSlivers() {
-    final dry = controller.dryRunResults;
-    final applied = controller.applyResults;
-    final slivers = <Widget>[];
-    if (dry != null) {
-      slivers
-        ..add(_gap(PlinkSpacing.s5))
-        ..addAll(_resultSectionSlivers(
-          title: 'Dry-run result',
-          subtitle: 'No changes were written. This is what an apply would do.',
-          results: dry,
-        ));
-    }
-    if (applied != null) {
-      slivers
-        ..add(_gap(PlinkSpacing.s5))
-        ..addAll(_resultSectionSlivers(
-          title: 'Apply result',
-          subtitle: 'Written to the target systems.',
-          results: applied,
-        ));
-    }
-    return slivers;
-  }
-
-  /// One dry-run/apply result set (#111): its header is an adapter and its
-  /// outcome rows render through a lazy [SliverList] — an apply over thousands
-  /// of accounts yields thousands of rows.
-  List<Widget> _resultSectionSlivers({
-    required String title,
-    required String subtitle,
-    required List<ActionOutcomeEntry> results,
-  }) =>
-      <Widget>[
-        _section(_ResultsHeader(title: title, subtitle: subtitle)),
-        SliverPadding(
-          padding: _hPad,
-          sliver: SliverList.builder(
-            itemCount: results.length,
-            itemBuilder: (context, index) => _ResultRow(result: results[index]),
-          ),
-        ),
-      ];
 }
 
 class _Header extends StatelessWidget {
@@ -631,736 +507,6 @@ class _DuplicateWarningTile extends StatelessWidget {
                     icon: const Icon(Icons.check, size: 16),
                     label: const Text('Deze dubbele mail accepteren'),
                   ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// Shows the apply-confirmation dialog and, on confirm, runs [apply] (#110).
-/// Shared by the global, per-situation, and per-entry apply affordances so a
-/// write is always one deliberate confirmation.
-Future<void> _confirmAndApply(
-  BuildContext context, {
-  required String title,
-  required int count,
-  required Future<void> Function() apply,
-}) async {
-  final confirmed = await showDialog<bool>(
-    context: context,
-    builder: (context) => AlertDialog(
-      title: Text(title),
-      content: Text(
-        'This writes $count change(s) to Smartschool and Azure AD. '
-        'Run a dry-run first to preview the exact changes.',
-      ),
-      actions: <Widget>[
-        TextButton(
-          onPressed: () => Navigator.of(context).pop(false),
-          child: const Text('Cancel'),
-        ),
-        FilledButton(
-          key: const ValueKey('reconcile-apply-confirm'),
-          onPressed: () => Navigator.of(context).pop(true),
-          child: const Text('Apply'),
-        ),
-      ],
-    ),
-  );
-  if (confirmed ?? false) await apply();
-}
-
-/// The pending-actions title (#110/#111): the "Pending actions (N)" heading,
-/// plus the "everything is in sync" line when there is nothing to do. Kept a
-/// fixed-size adapter above the lazy entry list.
-class _PendingActionsHeader extends StatelessWidget {
-  const _PendingActionsHeader({required this.count});
-
-  final int count;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Text('Pending actions ($count)', style: text.titleMedium),
-        const SizedBox(height: PlinkSpacing.s3),
-        if (count == 0)
-          Text(
-            'Everything is in sync — no pending actions.',
-            style: text.bodyMedium,
-          ),
-      ],
-    );
-  }
-}
-
-/// The global dry-run/apply affordances (#110): the secondary escape hatch that
-/// acts on every pending entry's chosen resolution. Sits below the lazy entry
-/// list as a fixed-size adapter (#111).
-class _PendingActionsFooter extends StatelessWidget {
-  const _PendingActionsFooter({required this.controller});
-
-  final ReconcileController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(top: PlinkSpacing.s3),
-      child: Wrap(
-        spacing: PlinkSpacing.s3,
-        children: <Widget>[
-          OutlinedButton.icon(
-            key: const ValueKey('reconcile-dry-run'),
-            onPressed: controller.busy || controller.applyableCount == 0
-                ? null
-                : controller.dryRun,
-            icon: const Icon(Icons.visibility_outlined),
-            label: const Text('Dry-run all'),
-          ),
-          TextButton.icon(
-            key: const ValueKey('reconcile-apply'),
-            onPressed: controller.busy || controller.applyableCount == 0
-                ? null
-                : () => _confirmAndApply(
-                      context,
-                      title: 'Apply pending actions?',
-                      count: controller.applyableCount,
-                      apply: controller.applyAll,
-                    ),
-            icon: const Icon(Icons.play_arrow_outlined),
-            label: const Text('Apply all'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// The bulk header of one "same situation" subset (#110): the "apply this
-/// resolution to all" affordance that honours each entry's own chosen
-/// alternative. Rendered as its own row in the lazy list (#111), only when more
-/// than one account shares the situation — a lone entry is resolved from its
-/// own tile.
-class _SituationHeader extends StatelessWidget {
-  const _SituationHeader({required this.controller, required this.entries});
-
-  final ReconcileController controller;
-  final List<PendingAccountEntry> entries;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final key = entries.first.situationKey;
-    final applyable = entries.where((e) => e.canApply).length;
-
-    return Padding(
-      padding: const EdgeInsets.only(
-        top: PlinkSpacing.s2,
-        bottom: PlinkSpacing.s2,
-      ),
-      child: Row(
-        children: <Widget>[
-          Expanded(
-            child: Text(
-              '${entries.first.situationLabel} — ${entries.length} '
-              'accounts in the same situation',
-              style: text.titleSmall,
-            ),
-          ),
-          const SizedBox(width: PlinkSpacing.s2),
-          OutlinedButton(
-            key: ValueKey('situation-dry-run-$key'),
-            onPressed: controller.busy || applyable == 0
-                ? null
-                : () => controller.dryRunSituation(key),
-            child: const Text('Dry-run all'),
-          ),
-          const SizedBox(width: PlinkSpacing.s2),
-          FilledButton(
-            key: ValueKey('situation-apply-$key'),
-            onPressed: controller.busy || applyable == 0
-                ? null
-                : () => _confirmAndApply(
-                      context,
-                      title: 'Apply to ${entries.length} accounts?',
-                      count: applyable,
-                      apply: () => controller.applySituation(key),
-                    ),
-            child: Text('Apply to all ($applyable)'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// One account's pending resolution (#110): a single expandable row showing the
-/// selected summary, the mutually-exclusive choice (as radios) when there is
-/// one, the per-field diff, and per-entry dry-run / apply.
-class _PendingEntryTile extends StatelessWidget {
-  const _PendingEntryTile({required this.controller, required this.entry});
-
-  final ReconcileController controller;
-  final PendingAccountEntry entry;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final Color hairline = Theme.of(context).dividerColor;
-
-    String lineFor(PendingChoice c) {
-      final summary = c.selected.changes.summary;
-      if (c.isChoice) return '$summary (keuze)';
-      return c.selected.canApply ? summary : '$summary (manueel)';
-    }
-
-    return Container(
-      margin: const EdgeInsets.only(bottom: PlinkSpacing.s2),
-      decoration: BoxDecoration(
-        border: Border.all(color: hairline),
-        borderRadius: const BorderRadius.all(Radius.circular(PlinkRadius.base)),
-      ),
-      child: ExpansionTile(
-        key: ValueKey('entry-${entry.family}-${entry.targetId}'),
-        shape: const Border(),
-        collapsedShape: const Border(),
-        leading: PlinkBadge(entry.family),
-        title: Text(entry.target, style: text.bodyLarge),
-        subtitle: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            for (final c in entry.choices)
-              Text(lineFor(c), style: text.bodySmall),
-          ],
-        ),
-        childrenPadding: const EdgeInsets.fromLTRB(
-          PlinkSpacing.s5,
-          0,
-          PlinkSpacing.s5,
-          PlinkSpacing.s4,
-        ),
-        expandedCrossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          for (final choice in entry.choices)
-            if (choice.isChoice)
-              _ChoiceControl(
-                controller: controller,
-                entry: entry,
-                choice: choice,
-              )
-            else
-              _OptionDetail(option: choice.selected),
-          const SizedBox(height: PlinkSpacing.s3),
-          Row(
-            children: <Widget>[
-              OutlinedButton(
-                key: ValueKey('entry-dry-run-${entry.targetId}'),
-                onPressed: controller.busy || !entry.canApply
-                    ? null
-                    : () => controller.dryRunEntry(entry),
-                child: const Text('Dry-run'),
-              ),
-              const SizedBox(width: PlinkSpacing.s2),
-              FilledButton(
-                key: ValueKey('entry-apply-${entry.targetId}'),
-                onPressed: controller.busy || !entry.canApply
-                    ? null
-                    : () => _confirmAndApply(
-                          context,
-                          title: 'Apply for ${entry.target}?',
-                          count: entry.choices
-                              .where((c) => c.selected.canApply)
-                              .length,
-                          apply: () => controller.applyEntry(entry),
-                        ),
-                child: const Text('Apply'),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// The radio group for a mutually-exclusive choice (#110): the operator picks
-/// exactly one resolution; the selected one is what an apply runs.
-class _ChoiceControl extends StatelessWidget {
-  const _ChoiceControl({
-    required this.controller,
-    required this.entry,
-    required this.choice,
-  });
-
-  final ReconcileController controller;
-  final PendingAccountEntry entry;
-  final PendingChoice choice;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final ColorScheme colors = Theme.of(context).colorScheme;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Text(
-          'Kies één oplossing:',
-          style: text.bodySmall?.copyWith(fontWeight: FontWeight.w600),
-        ),
-        for (final option in choice.alternatives)
-          InkWell(
-            key: ValueKey('alt-${entry.targetId}-${option.kind}'),
-            onTap: controller.busy
-                ? null
-                : () => controller.chooseAlternative(
-                      entry: entry,
-                      group: option.group!,
-                      kind: option.kind,
-                    ),
-            child: Padding(
-              padding: const EdgeInsets.symmetric(vertical: PlinkSpacing.s1),
-              child: Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Icon(
-                    option.kind == choice.selected.kind
-                        ? Icons.radio_button_checked
-                        : Icons.radio_button_unchecked,
-                    size: 18,
-                    color: option.kind == choice.selected.kind
-                        ? colors.primary
-                        : Theme.of(context).disabledColor,
-                  ),
-                  const SizedBox(width: PlinkSpacing.s2),
-                  Expanded(
-                    child: Text(option.changes.summary, style: text.bodySmall),
-                  ),
-                ],
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-}
-
-/// The per-field diff (or a lifecycle note) for a single, non-choice option.
-class _OptionDetail extends StatelessWidget {
-  const _OptionDetail({required this.option});
-
-  final PendingActionOption option;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final fields = option.changes.fields;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: fields.isEmpty
-          ? <Widget>[
-              Text(
-                option.canApply
-                    ? 'Lifecycle action — no per-field diff.'
-                    : '${option.changes.summary} '
-                        '(manual — not applied automatically)',
-                style: text.bodySmall,
-              ),
-            ]
-          : <Widget>[
-              for (final f in fields)
-                Padding(
-                  padding: const EdgeInsets.only(bottom: PlinkSpacing.s1),
-                  child: Text(
-                    '${f.field}: ${f.before ?? '∅'} → ${f.after ?? '∅'}',
-                    style: text.bodySmall,
-                  ),
-                ),
-            ],
-    );
-  }
-}
-
-/// The materialized overview (#115): the school → grade-year → classroom
-/// drill-down driven by the stored rollups, so it renders from the shared state
-/// even in a passive session that never pulled or re-linked.
-class _DrillDownSection extends StatelessWidget {
-  const _DrillDownSection({required this.controller});
-
-  final ReconcileController controller;
-
-  String? _freshness() {
-    final state = controller.syncState;
-    if (state.generation == 0) return null;
-    final at = state.updatedAt;
-    final when = at == null
-        ? ''
-        : ' · ${at.toLocal().hour.toString().padLeft(2, '0')}:'
-            '${at.toLocal().minute.toString().padLeft(2, '0')}';
-    final who = state.updatedBy == null || state.updatedBy!.isEmpty
-        ? ''
-        : ' door ${state.updatedBy}';
-    return 'Generatie ${state.generation}$when$who';
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final Color hairline = Theme.of(context).dividerColor;
-    final schools = controller.schoolRollups;
-    final groups = controller.groupRollup;
-    final freshness = _freshness();
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Text('Overzicht', style: text.titleMedium),
-        if (freshness != null) ...<Widget>[
-          const SizedBox(height: PlinkSpacing.s1),
-          Text(freshness, style: text.bodySmall),
-        ],
-        const SizedBox(height: PlinkSpacing.s3),
-        if (schools.isEmpty && groups == null)
-          Text('Nog geen gematerialiseerd overzicht.', style: text.bodyMedium)
-        else ...<Widget>[
-          for (final school in schools)
-            Container(
-              margin: const EdgeInsets.only(bottom: PlinkSpacing.s2),
-              decoration: BoxDecoration(
-                border: Border.all(color: hairline),
-                borderRadius:
-                    const BorderRadius.all(Radius.circular(PlinkRadius.base)),
-              ),
-              child: ExpansionTile(
-                key: ValueKey('rollup-school-${school.key}'),
-                shape: const Border(),
-                collapsedShape: const Border(),
-                title: Text(school.label, style: text.bodyLarge),
-                trailing: _PendingBadge(count: school.pendingCount),
-                children: <Widget>[
-                  for (final grade in controller.childrenOf(school.key))
-                    _GradeNode(controller: controller, grade: grade),
-                ],
-              ),
-            ),
-          // The group-family node (#119): class-group actions live outside the
-          // school → grade-year → classroom tree, so they get their own entry.
-          if (groups != null)
-            Container(
-              margin: const EdgeInsets.only(bottom: PlinkSpacing.s2),
-              decoration: BoxDecoration(
-                border: Border.all(color: hairline),
-                borderRadius:
-                    const BorderRadius.all(Radius.circular(PlinkRadius.base)),
-              ),
-              child: ListTile(
-                key: const ValueKey('rollup-groups'),
-                title: Text(groups.label, style: text.bodyLarge),
-                subtitle: Text('${groups.accountCount} klasgroep(en)',
-                    style: text.bodySmall),
-                trailing: _PendingBadge(count: groups.pendingCount),
-                onTap: controller.openGroups,
-              ),
-            ),
-        ],
-      ],
-    );
-  }
-}
-
-/// The "Klasgroepen" drill-down (#119): the per-group docs for the group-action
-/// family, lazily loaded from the store — the group counterpart of
-/// [_ClassroomDetail]. Renders even in a passive session that never linked.
-class _GroupsDetail extends StatelessWidget {
-  const _GroupsDetail({required this.controller});
-
-  final ReconcileController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final groups = controller.groupDocs;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Row(
-          children: <Widget>[
-            TextButton.icon(
-              key: const ValueKey('reconcile-groups-back'),
-              onPressed: controller.closeGroups,
-              icon: const Icon(Icons.arrow_back, size: 16),
-              label: const Text('Overzicht'),
-            ),
-            const SizedBox(width: PlinkSpacing.s2),
-            Text(controller.groupRollup?.label ?? 'Klasgroepen',
-                style: text.titleMedium),
-          ],
-        ),
-        const SizedBox(height: PlinkSpacing.s3),
-        if (controller.loadingGroups)
-          const LinearProgressIndicator()
-        else if (groups == null || groups.isEmpty)
-          Text('Geen klasgroepen met openstaande acties.',
-              style: text.bodyMedium)
-        else
-          for (final group in groups) _GroupTile(group: group),
-      ],
-    );
-  }
-}
-
-class _GroupTile extends StatelessWidget {
-  const _GroupTile({required this.group});
-
-  final MaterializedGroup group;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final Color hairline = Theme.of(context).dividerColor;
-    final systems = <String>[
-      if (group.inWisa) 'WISA',
-      if (group.inSmartschool) 'Smartschool',
-      if (group.inAzure) 'Azure',
-    ];
-
-    return Container(
-      margin: const EdgeInsets.only(bottom: PlinkSpacing.s2),
-      padding: const EdgeInsets.all(PlinkSpacing.s4),
-      decoration: BoxDecoration(
-        border: Border.all(color: hairline),
-        borderRadius: const BorderRadius.all(Radius.circular(PlinkRadius.base)),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Row(
-            children: <Widget>[
-              Expanded(child: Text(group.label, style: text.bodyLarge)),
-              _PendingBadge(
-                  count: group.candidates.where((c) => c.canApply).length),
-            ],
-          ),
-          const SizedBox(height: PlinkSpacing.s2),
-          Wrap(
-            spacing: PlinkSpacing.s2,
-            children: <Widget>[for (final s in systems) PlinkBadge(s)],
-          ),
-          for (final c in group.candidates)
-            Padding(
-              padding: const EdgeInsets.only(top: PlinkSpacing.s1),
-              child: Text(
-                c.canApply ? '• ${c.summary}' : '• ${c.summary} (manueel)',
-                style: text.bodySmall,
-              ),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-class _GradeNode extends StatelessWidget {
-  const _GradeNode({required this.controller, required this.grade});
-
-  final ReconcileController controller;
-  final Rollup grade;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    return ExpansionTile(
-      key: ValueKey('rollup-grade-${grade.key}'),
-      shape: const Border(),
-      collapsedShape: const Border(),
-      tilePadding: const EdgeInsets.only(left: PlinkSpacing.s5, right: 16),
-      title: Text('Jaar ${grade.label}', style: text.bodyMedium),
-      trailing: _PendingBadge(count: grade.pendingCount),
-      children: <Widget>[
-        for (final classroom in controller.childrenOf(grade.key))
-          ListTile(
-            key: ValueKey('rollup-class-${classroom.key}'),
-            contentPadding:
-                const EdgeInsets.only(left: PlinkSpacing.s6, right: 16),
-            title: Text(classroom.label, style: text.bodyMedium),
-            subtitle: Text('${classroom.accountCount} account(s)',
-                style: text.bodySmall),
-            trailing: _PendingBadge(count: classroom.pendingCount),
-            onTap: () => controller.openClassroom(classroom),
-          ),
-      ],
-    );
-  }
-}
-
-class _PendingBadge extends StatelessWidget {
-  const _PendingBadge({required this.count});
-
-  final int count;
-
-  @override
-  Widget build(BuildContext context) {
-    if (count == 0) {
-      return Icon(Icons.check,
-          size: 16, color: Theme.of(context).disabledColor);
-    }
-    return PlinkBadge('$count');
-  }
-}
-
-/// One classroom's accounts, lazily loaded from the store on drill-down.
-class _ClassroomDetail extends StatelessWidget {
-  const _ClassroomDetail({required this.controller});
-
-  final ReconcileController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final classroom = controller.selectedClassroom;
-    final accounts = controller.classroomAccounts;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Row(
-          children: <Widget>[
-            TextButton.icon(
-              key: const ValueKey('reconcile-classroom-back'),
-              onPressed: controller.closeClassroom,
-              icon: const Icon(Icons.arrow_back, size: 16),
-              label: const Text('Overzicht'),
-            ),
-            const SizedBox(width: PlinkSpacing.s2),
-            Text(classroom?.label ?? '', style: text.titleMedium),
-          ],
-        ),
-        const SizedBox(height: PlinkSpacing.s3),
-        if (controller.loadingClassroom)
-          const LinearProgressIndicator()
-        else if (accounts == null || accounts.isEmpty)
-          Text('Geen accounts in deze klas.', style: text.bodyMedium)
-        else
-          for (final account in accounts) _AccountTile(account: account),
-      ],
-    );
-  }
-}
-
-class _AccountTile extends StatelessWidget {
-  const _AccountTile({required this.account});
-
-  final MaterializedAccount account;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final Color hairline = Theme.of(context).dividerColor;
-    final systems = <String>[
-      if (account.inWisa) 'WISA',
-      if (account.inSmartschool) 'Smartschool',
-      if (account.inAzure) 'Azure',
-    ];
-
-    return Container(
-      margin: const EdgeInsets.only(bottom: PlinkSpacing.s2),
-      padding: const EdgeInsets.all(PlinkSpacing.s4),
-      decoration: BoxDecoration(
-        border: Border.all(color: hairline),
-        borderRadius: const BorderRadius.all(Radius.circular(PlinkRadius.base)),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Row(
-            children: <Widget>[
-              Expanded(child: Text(account.label, style: text.bodyLarge)),
-              _PendingBadge(
-                  count: account.candidates.where((c) => c.canApply).length),
-            ],
-          ),
-          const SizedBox(height: PlinkSpacing.s2),
-          Wrap(
-            spacing: PlinkSpacing.s2,
-            children: <Widget>[for (final s in systems) PlinkBadge(s)],
-          ),
-          for (final w in account.warnings)
-            Padding(
-              padding: const EdgeInsets.only(top: PlinkSpacing.s1),
-              child: Text(w, style: text.bodySmall),
-            ),
-          for (final c in account.candidates)
-            Padding(
-              padding: const EdgeInsets.only(top: PlinkSpacing.s1),
-              child: Text('• ${c.summary}', style: text.bodySmall),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-/// The header of a dry-run/apply result set: its title and subtitle, above the
-/// lazy list of outcome rows (#111).
-class _ResultsHeader extends StatelessWidget {
-  const _ResultsHeader({required this.title, required this.subtitle});
-
-  final String title;
-  final String subtitle;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Text(title, style: text.titleMedium),
-        const SizedBox(height: PlinkSpacing.s1),
-        Text(subtitle, style: text.bodySmall),
-        const SizedBox(height: PlinkSpacing.s3),
-      ],
-    );
-  }
-}
-
-/// One outcome row of a dry-run/apply pass (#111): the check/cross plus the
-/// target and its change summary (or the failure cause). Built lazily by the
-/// results [SliverList].
-class _ResultRow extends StatelessWidget {
-  const _ResultRow({required this.result});
-
-  final ActionOutcomeEntry result;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme text = Theme.of(context).textTheme;
-    final ColorScheme colors = Theme.of(context).colorScheme;
-    final failed = result.outcome == actions.ActionOutcome.failed;
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: PlinkSpacing.s1),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          Icon(
-            failed ? Icons.close : Icons.check,
-            size: 16,
-            color: failed ? colors.error : colors.primary,
-          ),
-          const SizedBox(width: PlinkSpacing.s2),
-          Expanded(
-            child: Text(
-              failed
-                  ? '${result.target} — ${result.changes.summary}: '
-                      '${result.error}'
-                  : '${result.target} — ${result.changes.summary}',
-              style: text.bodySmall,
-            ),
           ),
         ],
       ),
